Tentative fixes for:
1. Bug #13, adds support for [BelongsTo(foreignKey="*")] syntax. Thanks
Rene!
2. Bug #12, wrapped a few things in processNtoNRelationships in a
try/catch block. If any of the dependencies have not been fetched then
N-N relationship processing may fail. This will be logged (WARN) but
won't blow up with an exception.
3. cleanupModelReferences have been moved from update call to
unmarshallNode block. It should only be invoked if the references have
actually been changed. Not a good idea to just pro-actively call it on
update.

Added Ruboss.filters and Ruboss.filters$ shortcut functions that allow
application of filters arrays on RubossCollections.